test: extract shared @flareapp/test-helpers (FakeApi, fakeSpan)
Scaffolds the private @flareapp/test-helpers workspace package and moves
the duplicated FakeApi and fakeSpan fixtures from core/js/react-native
into it. Old tests/helpers files become re-export shims so the ~26
existing consumer imports (./helpers and ./helpers/FakeApi) don't change.

browserTracing.test.ts and idleRootController.test.ts keep their own
local fakeSpan variants (parameterized id/traceId/endTime, vi.fn-backed
end() inspected via toHaveBeenCalledWith) since the shared fakeSpan's
shape is incompatible with those assertions without changing them.
