+++ packages/js/tests/helpers/FakeApi.ts
@@ -1,20 +1,1 @@
-import { Api } from '../../src/api';
-import type { Report } from '../../src/types';
-
-export class FakeApi extends Api {
-    reports: Report[] = [];
-
-    lastReport?: Report;
-    lastUrl?: string;
-    lastKey?: string | null;
-    lastReportBrowserExtensionErrors?: boolean;
-
-    report(report: Report, url: string, key: string | null, reportBrowserExtensionErrors: boolean): Promise<void> {
-        this.reports.push(report);
-        this.lastUrl = url;
-        this.lastKey = key;
-        this.lastReportBrowserExtensionErrors = reportBrowserExtensionErrors;
-        this.lastReport = report;
-        return Promise.resolve();
-    }
-}
+export { FakeApi } from '@flareapp/test-helpers';
