+++ packages/js/tests/helpers/fakeTracer.ts
@@ -1,33 +1,10 @@
-import type { Config, Span, SpanOptions } from '@flareapp/core';
+import type { Config, SpanOptions } from '@flareapp/core';
+import { fakeSpan } from '@flareapp/test-helpers';
 import { vi } from 'vitest';
 
 import type { HttpTracer } from '../../src/tracing/httpRequestSpan';
 
-export function fakeSpan() {
-    const calls = { attrs: {} as Record<string, unknown>, status: undefined as unknown, ended: false };
-    const span: Span = {
-        traceId: 'a'.repeat(32),
-        spanId: 'b'.repeat(16),
-        parentSpanId: null,
-        name: '',
-        isRecording: true,
-        setAttribute(k, v) {
-            calls.attrs[k] = v;
-            return this;
-        },
-        setStatus(s) {
-            calls.status = s;
-            return this;
-        },
-        addEvent() {
-            return this;
-        },
-        end() {
-            calls.ended = true;
-        },
-    };
-    return { span, calls };
-}
+export { fakeSpan };
 
 /**
  * `startSpan` creates a fresh fake span per call (each with its own `calls`), pushing every one's
