+++ packages/react-native/package.json
@@ -56,6 +56,7 @@
     "devDependencies": {
         "@flareapp/js": "file:../js",
         "@flareapp/react": "file:../react",
+        "@flareapp/test-helpers": "*",
         "@types/react": "^19.0.0",
         "react": "^19.0.0",
         "react-native": "^0.79.0",
