+++ packages/test-helpers/package.json
@@ -1,0 +1,18 @@
+{
+    "name": "@flareapp/test-helpers",
+    "private": true,
+    "type": "module",
+    "main": "./src/index.ts",
+    "types": "./src/index.ts",
+    "exports": {
+        ".": "./src/index.ts"
+    },
+    "scripts": {
+        "typescript": "tsc --noEmit"
+    },
+    "devDependencies": {
+        "@flareapp/core": "*",
+        "typescript": "^5.7.0",
+        "vitest": "^4.0.0"
+    }
+}
